backend/requirements.txt
django
djangorestframework
djangorestframework-jwt
django-cors-headers
psycopg2-binary
